pyproject.toml
[tool.poetry]
name = "doc-to-sent"
version = "0.1.0"
description = ""
authors = ["David Baines <david_baines@sil.org>"]
readme = "README.md"
packages = [{include = "doc_to_sent"}]

[tool.poetry.dependencies]
python = "^3.8"
python-docx = "^0.8.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
